SCoin_count paramater change to float. Temporary money button removed.

+++ src/GUI/DeveloperButton.java
@@ -212,7 +212,7 @@ public class DeveloperButton extends JButton
         return LogoImagePath;
     }
 
-    public void setEnabledByScoin(int SCoin)
+    public void setEnabledByScoin(float SCoin)
     {
         this.setEnabled(Employee.getTotalPrice(LOC.buy_amount) <= SCoin);
     }
